init: initial setup
For use on a nixOS virtual machine with hostname 'test-vm'.
Note, that the password login is encrypted. Before use in a new
environment, ensure that the './sops.yaml' contains the age
encrypted public key 'ssh_host_ed25519_key.pub', the 'secrets.yaml'
has been updated with sops and the 'hardware-configuration.nix' has
been replaced with your local install.

At this stage the only secret needed for a successful run is
'user-roman-hashedpwd'.